requirements.txt
asgiref==3.2.3
cffi==1.13.2
cryptography==2.8
Django==2.1.15
django-cors-headers==3.2.0
django-rest-auth==0.9.5
djangorestframework==3.10.3
mysqlclient==1.4.6
pycparser==2.19
PyMySQL==0.9.3
python-dotenv==0.10.3
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
